+++ CHANGELOG.md
@@ -32,6 +32,9 @@
 
 
 
+* Datastream
+  * Support for `S3-Compatible`, `TrafficPeak`, and `Dynatrace` destinations has been added.
+  * Migrated Datastream APIs from `V2` to `V3` for `CDN` log types. Other log types available in `V3` APIs (`eDNS`, `GTM` and `Edgeworkers`) are not yet supported.